third_party/WebKit/LayoutTests/inspector/extensions/extensions-panel-expected.txt

CONSOLE MESSAGE: line 12: hello
Tests WebInspector extension API

Started extension.
Running tests...
RUNNING TEST: extension_testCreatePanel
Panel created
{
    createStatusBarButton : <function>
    onHidden : {
        addListener : <function>
        removeListener : <function>
    }
    onSearch : {
        addListener : <function>
        removeListener : <function>
    }
    onShown : {
        addListener : <function>
        removeListener : <function>
    }
    show : <function>
}
Panel shown
Extension panel size correct
Panel hidden
RUNNING TEST: extension_testGlobalShortcuts
Panel shown, now toggling console...
Panel resized, switching away...
Panel hidden, test passed.
RUNNING TEST: extension_testOpenResource
Showing .../extensions-panel.html
Showing resource .../inspector/extensions/extensions-panel.html in panel sources (sources), line: 1001
Showing .../abe.png
Showing resource .../extensions/resources/abe.png in panel resources (resources), line: 1002
Showing .../missing.txt
Showing resource .../extensions/resources/missing.txt in panel network (network), line: undefined
Showing not-found.html
log: Extension server error: Object not found: not-found.html
Showing javascript:console.error('oh no!')
log: Extension server error: Object not found: javascript:console.error('oh no!')
RUNNING TEST: extension_testSearch
Panel searched:
{
    0 : "performSearch"
    1 : "hello"
}
Panel searched:
{
    0 : "cancelSearch"
    1 : undefined
}
RUNNING TEST: extension_testStatusBarButtons
Created a status bar button, dump follows:
{
    onClicked : {
        addListener : <function>
        removeListener : <function>
    }
    update : <function>
}
button1 clicked
Status bar buttons state:
status bar item 0, icon: .../inspector/resources/button1.png, tooltip: 'Button One tooltip', disabled: false
status bar item 1, icon: .../inspector/resources/button2.png, tooltip: 'Button Two tooltip', disabled: true
Updated status bar buttons
Status bar buttons state:
status bar item 0, icon: .../inspector/resources/button1-updated.png, tooltip: 'Button One tooltip', disabled: false
status bar item 1, icon: .../inspector/resources/button2.png, tooltip: 'Button Two updated tooltip', disabled: false
button2 clicked
All tests done.
